new size must be int (duh!)

+++ pyVideoGrabber/src/rstpGrabber.py
@@ -37,10 +37,10 @@ class RTSP_grabber(threading.Thread):
     def image_encoder(self, img):
         if img.width > 720:
             factor = 720/img.width
-            newWidth = img.width * factor
+            newWidth = int(img.width * factor)
             if newWidth % 2 != 0:
                 newWidth = newWidth + 1
-            newHeight = img.height * factor
+            newHeight = int(img.height * factor)
             if newHeight % 2 != 0:
                 newHeight = newHeight + 1
             newSize = (newWidth, newHeight)
